requirements.txt
streamlit
python-dotenv
langdetect
nltk
langchain
langchain-google-genai
uuid
fuzzywuzzy
python-Levenshtein